refactor slices
