--- app/javascript/reducers/index.js
@@ -1,8 +1,0 @@
-import { combineReducers } from '@reduxjs/toolkit';
-
-import editorReducer from '../features/Editor/editorSlice.js';
-
-
-export const rootReducer = combineReducers({
-  editor: editorReducer,
-});
