+++ app/javascript/slices/index.js
@@ -1,0 +1,10 @@
+import { combineReducers } from '@reduxjs/toolkit';
+
+import editor from './editor.js';
+import terminal from './terminal.js';
+
+
+export default combineReducers({
+  editor,
+  terminal,
+});
